+++ twitter_back/api/urls.py
@@ -5,7 +5,10 @@ from api import views
 
 urlpatterns = [
     # path("", views.index, name="index"),
-    path("login/", obtain_jwt_token),
-    path("users/", views.UserListAPIView.as_view()),
-    path("users/<int:id>/", views.UserDetailAPIView.as_view())
+    # path("login/", obtain_jwt_token),
+    # path("users/", views.UserListAPIView.as_view()),
+    # path("users/<int:id>/", views.UserDetailAPIView.as_view())
+    path("news/", views.posts_list),
+    path("news/<int:pk>/", views.post_retrieve)
+
 ]
