+++ twitter_back/api/views/__init__.py
@@ -1,1 +1,2 @@
 from .cbv1 import UserListAPIView, UserDetailAPIView
+from .PostViews import posts_list, post_retrieve
